implemented TagInfos for the boosted double-b tagger
