+++ DataFormats/BTauReco/interface/TaggingVariable.h
@@ -125,6 +125,34 @@ namespace reco {
       leptonRatioRel,                           // momentum of the soft lepton parallel to jet axis over jet energy
       electronMVA,                              // mva output from electron ID
 
+      // ### specific to boosted double-b tagger (see BTV-15-002 PAS for more details) ###
+      trackSip3dSig_0,                          // 1st largest track 3D signed impact parameter significance
+      trackSip3dSig_1,                          // 2nd largest track 3D signed impact parameter significance
+      trackSip3dSig_2,                          // 3rd largest track 3D signed impact parameter significance
+      trackSip3dSig_3,                          // 4th largest track 3D signed impact parameter significance
+      tau1_trackSip3dSig_0,                     // 1st largest track 3D signed impact parameter significance associated to the 1st N-subjettiness axis
+      tau1_trackSip3dSig_1,                     // 2nd largest track 3D signed impact parameter significance associated to the 1st N-subjettiness axis
+      tau2_trackSip3dSig_0,                     // 1st largest track 3D signed impact parameter significance associated to the 2nd N-subjettiness axis
+      tau2_trackSip3dSig_1,                     // 2nd largest track 3D signed impact parameter significance associated to the 2nd N-subjettiness axis
+      trackSip2dSigAboveBottom_0,               // track 2D signed impact parameter significance of 1st track lifting mass above bottom
+      trackSip2dSigAboveBottom_1,               // track 2D signed impact parameter significance of 2nd track lifting mass above bottom
+      tau1_trackEtaRel_0,                       // 1st smallest track pseudorapidity, relative to the jet axis, associated to the 1st N-subjettiness axis
+      tau1_trackEtaRel_1,                       // 2nd smallest track pseudorapidity, relative to the jet axis, associated to the 1st N-subjettiness axis
+      tau1_trackEtaRel_2,                       // 3rd smallest track pseudorapidity, relative to the jet axis, associated to the 1st N-subjettiness axis
+      tau2_trackEtaRel_0,                       // 1st smallest track pseudorapidity, relative to the jet axis, associated to the 2nd N-subjettiness axis
+      tau2_trackEtaRel_1,                       // 2nd smallest track pseudorapidity, relative to the jet axis, associated to the 2nd N-subjettiness axis
+      tau2_trackEtaRel_2,                       // 3rd smallest track pseudorapidity, relative to the jet axis, associated to the 2nd N-subjettiness axis
+      tau1_vertexMass,                          // mass of track sum at secondary vertex associated to the 1st N-subjettiness axis
+      tau1_vertexEnergyRatio,                   // ratio of energy at secondary vertex over total energy associated to the 1st N-subjettiness axis
+      tau1_flightDistance2dSig,                 // transverse distance significance between primary and secondary vertex associated to the 1st N-subjettiness axis
+      tau1_vertexDeltaR,                        // pseudoangular distance between the 1st N-subjettiness axis and secondary vertex direction
+      tau2_vertexMass,                          // mass of track sum at secondary vertex associated to the 2nd N-subjettiness axis
+      tau2_vertexEnergyRatio,                   // ratio of energy at secondary vertex over total energy associated to the 2nd N-subjettiness axis
+      tau2_flightDistance2dSig,                 // transverse distance significance between primary and secondary vertex associated to the 2nd N-subjettiness axis
+      tau2_vertexDeltaR,                        // pseudoangular distance between the 2nd N-subjettiness axis and secondary vertex direction
+      z_ratio,                                  // z ratio
+      // #################################################################################
+
       algoDiscriminator,                        // discriminator output of an algorithm
 
       lastTaggingVariable
