+++ DataFormats/BTauReco/src/TaggingVariable.cc
@@ -99,10 +99,35 @@ const char* const TaggingVariableDescription[] = {
   /* [leptonP0Par]                              = */ "momentum of the soft lepton along the jet direction, in the jet rest frame",
   /* [leptonEtaRel]                             = */ "pseudo)rapidity of the soft lepton along jet axis",
   /* [leptonDeltaR]                             = */ "pseudo)angular distance of the soft lepton to jet axis",
-  /* [leptonRatio],                             = */ "momentum of the soft lepton over jet energy",
+  /* [leptonRatio]                              = */ "momentum of the soft lepton over jet energy",
   /* [leptonRatioRel]                           = */ "momentum of the soft lepton parallel to jet axis over jet energy",
   /* [electronMVA]                              = */ "mva output of the electron ID",
 
+  /* [trackSip3dSig_0]                          = */ "1st largest track 3D signed impact parameter significance",
+  /* [trackSip3dSig_1]                          = */ "2nd largest track 3D signed impact parameter significance",
+  /* [trackSip3dSig_2]                          = */ "3rd largest track 3D signed impact parameter significance",
+  /* [trackSip3dSig_3]                          = */ "4th largest track 3D signed impact parameter significance",
+  /* [tau1_trackSip3dSig_0]                     = */ "1st largest track 3D signed impact parameter significance associated to the 1st N-subjettiness axis",
+  /* [tau1_trackSip3dSig_1]                     = */ "2nd largest track 3D signed impact parameter significance associated to the 1st N-subjettiness axis",
+  /* [tau2_trackSip3dSig_0]                     = */ "1st largest track 3D signed impact parameter significance associated to the 2nd N-subjettiness axis",
+  /* [tau2_trackSip3dSig_1]                     = */ "2nd largest track 3D signed impact parameter significance associated to the 2nd N-subjettiness axis",
+  /* [trackSip2dSigAboveBottom_0]               = */ "track 2D signed impact parameter significance of 1st track lifting mass above bottom",
+  /* [trackSip2dSigAboveBottom_1]               = */ "track 2D signed impact parameter significance of 2nd track lifting mass above bottom",
+  /* [tau1_trackEtaRel_0]                       = */ "1st smallest track pseudorapidity, relative to the jet axis, associated to the 1st N-subjettiness axis",
+  /* [tau1_trackEtaRel_1]                       = */ "2nd smallest track pseudorapidity, relative to the jet axis, associated to the 1st N-subjettiness axis",
+  /* [tau1_trackEtaRel_2]                       = */ "3rd smallest track pseudorapidity, relative to the jet axis, associated to the 1st N-subjettiness axis",
+  /* [tau2_trackEtaRel_0]                       = */ "1st smallest track pseudorapidity, relative to the jet axis, associated to the 2nd N-subjettiness axis",
+  /* [tau2_trackEtaRel_1]                       = */ "2nd smallest track pseudorapidity, relative to the jet axis, associated to the 2nd N-subjettiness axis",
+  /* [tau2_trackEtaRel_2]                       = */ "3rd smallest track pseudorapidity, relative to the jet axis, associated to the 2nd N-subjettiness axis",
+  /* [tau1_vertexMass]                          = */ "mass of track sum at secondary vertex associated to the 1st N-subjettiness axis",
+  /* [tau1_vertexEnergyRatio]                   = */ "ratio of energy at secondary vertex over total energy associated to the 1st N-subjettiness axis",
+  /* [tau1_flightDistance2dSig]                 = */ "transverse distance significance between primary and secondary vertex associated to the 1st N-subjettiness axis",
+  /* [tau1_vertexDeltaR]                        = */ "pseudoangular distance between the 1st N-subjettiness axis and secondary vertex direction",
+  /* [tau2_vertexMass]                          = */ "mass of track sum at secondary vertex associated to the 2nd N-subjettiness axis",
+  /* [tau2_vertexEnergyRatio]                   = */ "ratio of energy at secondary vertex over total energy associated to the 2nd N-subjettiness axis",
+  /* [tau2_flightDistance2dSig]                 = */ "transverse distance significance between primary and secondary vertex associated to the 2nd N-subjettiness axis",
+  /* [tau2_vertexDeltaR]                        = */ "pseudoangular distance between the 2nd N-subjettiness axis and secondary vertex direction",
+  /* [z_ratio]                                  = */ "z ratio",
 
   /* [algoDiscriminator]                        = */ "discriminator output of an algorithm",
 
@@ -182,19 +207,19 @@ const char* const TaggingVariableTokens[] = {
   /* [neutralHadronEnergyFraction]              = */ "neutralHadronEnergyFraction",
   /* [photonEnergyFraction]                     = */ "photonEnergyFraction",
   /* [electronEnergyFraction]                   = */ "electronEnergyFraction",
-  /* [muonEnergyFraction],                      = */ "muonEnergyFraction",
-  /* [chargedHadronMultiplicity],               = */ "chargedHadronMultiplicity",
-  /* [neutralHadronMultiplicity],               = */ "neutralHadronMultiplicity",
+  /* [muonEnergyFraction]                       = */ "muonEnergyFraction",
+  /* [chargedHadronMultiplicity]                = */ "chargedHadronMultiplicity",
+  /* [neutralHadronMultiplicity]                = */ "neutralHadronMultiplicity",
   /* [photonMultiplicity]                       = */ "photonMultiplicity",
   /* [electronMultiplicity]                     = */ "electronMultiplicity",
-  /* [muonMultiplicity],                        = */ "muonMultiplicity",
-  /* [hadronMultiplicity],                      = */ "hadronMultiplicity",
-  /* [hadronPhotonMultiplicity],                = */ "hadronPhotonMultiplicity",
+  /* [muonMultiplicity]                         = */ "muonMultiplicity",
+  /* [hadronMultiplicity]                       = */ "hadronMultiplicity",
+  /* [hadronPhotonMultiplicity]                 = */ "hadronPhotonMultiplicity",
   /* [totalMultiplicity]                        = */ "totalMultiplicity",
 
-  /* [massVertexEnergyFraction],                = */ "massVertexEnergyFraction",
-  /* [vertexBoostOverSqrtJetPt],                = */ "vertexBoostOverSqrtJetPt",
- 
+  /* [massVertexEnergyFraction]                 = */ "massVertexEnergyFraction",
+  /* [vertexBoostOverSqrtJetPt]                 = */ "vertexBoostOverSqrtJetPt",
+
   /* [leptonSip2d]                              = */ "leptonSip2d",
   /* [leptonSip3d]                              = */ "leptonSip3d",
   /* [leptonPtRel]                              = */ "leptonPtRel",
@@ -202,8 +227,34 @@ const char* const TaggingVariableTokens[] = {
   /* [leptonEtaRel]                             = */ "leptonEtaRel",
   /* [leptonDeltaR]                             = */ "leptonDeltaR",
   /* [leptonRatio]                              = */ "leptonRatio",
-  /* [leptonRatioRel],                          = */ "leptonRatioRel",
+  /* [leptonRatioRel]                           = */ "leptonRatioRel",
   /* [electronMVA]                              = */ "electronMVA",
+
+  /* [trackSip3dSig_0]                          = */ "trackSip3dSig_0",
+  /* [trackSip3dSig_1]                          = */ "trackSip3dSig_1",
+  /* [trackSip3dSig_2]                          = */ "trackSip3dSig_2",
+  /* [trackSip3dSig_3]                          = */ "trackSip3dSig_3",
+  /* [tau1_trackSip3dSig_0]                     = */ "tau1_trackSip3dSig_0",
+  /* [tau1_trackSip3dSig_1]                     = */ "tau1_trackSip3dSig_1",
+  /* [tau2_trackSip3dSig_0]                     = */ "tau2_trackSip3dSig_0",
+  /* [tau2_trackSip3dSig_1]                     = */ "tau2_trackSip3dSig_1",
+  /* [trackSip2dSigAboveBottom_0]               = */ "trackSip2dSigAboveBottom_0",
+  /* [trackSip2dSigAboveBottom_1]               = */ "trackSip2dSigAboveBottom_1",
+  /* [tau1_trackEtaRel_0]                       = */ "tau1_trackEtaRel_0",
+  /* [tau1_trackEtaRel_1]                       = */ "tau1_trackEtaRel_1",
+  /* [tau1_trackEtaRel_2]                       = */ "tau1_trackEtaRel_2",
+  /* [tau2_trackEtaRel_0]                       = */ "tau2_trackEtaRel_0",
+  /* [tau2_trackEtaRel_1]                       = */ "tau2_trackEtaRel_1",
+  /* [tau2_trackEtaRel_2]                       = */ "tau2_trackEtaRel_2",
+  /* [tau1_vertexMass]                          = */ "tau1_vertexMass",
+  /* [tau1_vertexEnergyRatio]                   = */ "tau1_vertexEnergyRatio",
+  /* [tau1_flightDistance2dSig]                 = */ "tau1_flightDistance2dSig",
+  /* [tau1_vertexDeltaR]                        = */ "tau1_vertexDeltaR",
+  /* [tau2_vertexMass]                          = */ "tau2_vertexMass",
+  /* [tau2_vertexEnergyRatio]                   = */ "tau2_vertexEnergyRatio",
+  /* [tau2_flightDistance2dSig]                 = */ "tau2_flightDistance2dSig",
+  /* [tau2_vertexDeltaR]                        = */ "tau2_vertexDeltaR",
+  /* [z_ratio]                                  = */ "z_ratio",
 
   /* [algoDiscriminator]                        = */ "algoDiscriminator",
