+++ DataFormats/BTauReco/src/classes.h
@@ -32,6 +32,7 @@
 #include "DataFormats/BTauReco/interface/CombinedTauTagInfo.h"
 #include "DataFormats/BTauReco/interface/SecondaryVertexTagInfo.h"
 #include "DataFormats/BTauReco/interface/CandSecondaryVertexTagInfo.h"
+#include "DataFormats/BTauReco/interface/BoostedDoubleSVTagInfo.h"
 #include "DataFormats/BTauReco/interface/SoftLeptonTagInfo.h"
 #include "DataFormats/BTauReco/interface/CandSoftLeptonTagInfo.h"
 #include "DataFormats/BTauReco/interface/TauImpactParameterInfo.h"
@@ -78,6 +79,13 @@ namespace DataFormats_BTauReco {
     reco::CandSecondaryVertexTagInfoRefVector                           casv_rv;
     edm::Wrapper<reco::CandSecondaryVertexTagInfoCollection>            casv_wc;
 
+    reco::BoostedDoubleSVTagInfo                                        bdsv;
+    reco::BoostedDoubleSVTagInfoCollection                              bdsv_c;
+    reco::BoostedDoubleSVTagInfoRef                                     bdsv_r;
+    reco::BoostedDoubleSVTagInfoFwdRef                                  bdsv_fr;
+    reco::BoostedDoubleSVTagInfoRefProd                                 bdsv_rp;
+    reco::BoostedDoubleSVTagInfoRefVector                               bdsv_rv;
+    edm::Wrapper<reco::BoostedDoubleSVTagInfoCollection>                bdsv_wc;
 
     reco::CombinedTauTagInfo                                            ct;
     reco::CombinedTauTagInfoCollection                                  ct_c;
@@ -273,6 +281,8 @@ namespace DataFormats_BTauReco {
     edm::reftobase::RefHolder<reco::SecondaryVertexTagInfoRef>                  rbh_sv;
     edm::reftobase::Holder<reco::BaseTagInfo, reco::CandSecondaryVertexTagInfoRef>  rb_casv;
     edm::reftobase::RefHolder<reco::CandSecondaryVertexTagInfoRef>                  rbh_casv;
+    edm::reftobase::Holder<reco::BaseTagInfo, reco::BoostedDoubleSVTagInfoRef>  rb_bdsv;
+    edm::reftobase::RefHolder<reco::BoostedDoubleSVTagInfoRef>                  rbh_bdsv;
     edm::reftobase::Holder<reco::BaseTagInfo, reco::CombinedTauTagInfoRef>      rb_ct;
     edm::reftobase::RefHolder<reco::CombinedTauTagInfoRef>                      rbh_ct;
     edm::reftobase::Holder<reco::BaseTagInfo, reco::IsolatedTauTagInfoRef>      rb_it;
@@ -302,6 +312,8 @@ namespace DataFormats_BTauReco {
     edm::reftobase::RefHolder<reco::SecondaryVertexTagInfoFwdRef>                  rbh_svf;
     edm::reftobase::Holder<reco::BaseTagInfo, reco::CandSecondaryVertexTagInfoFwdRef>  rb_casvf;
     edm::reftobase::RefHolder<reco::CandSecondaryVertexTagInfoFwdRef>                  rbh_casvf;
+    edm::reftobase::Holder<reco::BaseTagInfo, reco::BoostedDoubleSVTagInfoFwdRef>  rb_bdsvf;
+    edm::reftobase::RefHolder<reco::BoostedDoubleSVTagInfoFwdRef>                  rbh_bdsvf;
     edm::reftobase::Holder<reco::BaseTagInfo, reco::CombinedTauTagInfoFwdRef>      rb_ctf;
     edm::reftobase::RefHolder<reco::CombinedTauTagInfoFwdRef>                      rbh_ctf;
     edm::reftobase::Holder<reco::BaseTagInfo, reco::IsolatedTauTagInfoFwdRef>      rb_itf;
