+++ RecoBTag/SecondaryVertex/python/pfBoostedDoubleSecondaryVertexAK8BJetTags_cfi.py
@@ -2,6 +2,5 @@ import FWCore.ParameterSet.Config as cms
 
 pfBoostedDoubleSecondaryVertexAK8BJetTags = cms.EDProducer("JetTagProducer",
     jetTagComputer = cms.string('candidateBoostedDoubleSecondaryVertexAK8Computer'),
-    tagInfos = cms.VInputTag(cms.InputTag("pfImpactParameterAK8TagInfos"),
-                             cms.InputTag("pfInclusiveSecondaryVertexFinderAK8TagInfos"))
+    tagInfos = cms.VInputTag(cms.InputTag("pfBoostedDoubleSVAK8TagInfos"))
 )
