+++ RecoBTau/JetTagComputer/BuildFile.xml
@@ -8,7 +8,6 @@
 <use   name="FWCore/ParameterSet"/>
 <use   name="FWCore/Utilities"/>
 <use   name="PhysicsTools/MVAComputer"/>
-<use   name="TrackingTools/Records"/>
 <use   name="boost"/>
 <export>
   <lib   name="1"/>
